src/notion_poller/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "notion-network-sync"
version = "0.1.0"
description = "LinkedIn to Notion networking data poller"
authors = [{name = "Ethan McFarland"}]
requires-python = ">=3.8"
dependencies = [
    "notion-client",
    "python-dotenv", 
    "azure-ai-inference",
    "azure-core",
    "supabase",
] 

[tool.setuptools]
packages = ["common", "notion_poller"]
package-dir = {"" = "src"}